robotic disease, unullspace organs, remove status variable
https://github.com/tgstation/tgstation/pull/79687 https://github.com/tgstation/tgstation/pull/78519 https://github.com/tgstation/tgstation/pull/76350 https://github.com/tgstation/tgstation/pull/76766

+++ code/modules/mob/living/living.dm
@@ -1283,7 +1283,7 @@
 		Robot.notify_ai(NEW_BORG)
 	else
 		for(var/obj/item/item in src)
-			if(!dropItemToGround(item))
+			if(!dropItemToGround(item) && !(item.item_flags & ABSTRACT))
 				qdel(item)
 				continue
 			item_contents += item
